Refactor plugin to use WebSharks plugin format
